requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
tqdm>=4.66.4
faiss-cpu>=1.8.0
langchain>=0.2.10
langchain-community>=0.2.9
langchain-openai>=0.1.17
unstructured>=0.15.9
unstructured[pptx]>=0.15.9
pypdf>=4.2.0
python-docx>=1.1.2
atlassian-python-api>=3.41.16
